backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-binance==1.0.19
python-dotenv==1.0.0
websockets==12.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-jose==3.3.0
pandas==2.1.3
numpy==1.26.2
ta==0.10.2  # Technical Analysis library 